+++ packages/calcite-components/src/components/combobox/combobox.tsx
@@ -12,7 +12,7 @@ import {
   VNode,
   Watch,
 } from "@stencil/core";
-import { debounce } from "lodash-es";
+import { debounce, escapeRegExp } from "lodash-es";
 import { calciteSize48 } from "@esri/calcite-design-tokens/dist/es6/core.js";
 import { filter } from "../../utils/filter";
 import { getElementWidth, getTextWidth, toAriaBoolean } from "../../utils/dom";
@@ -115,6 +115,17 @@ export class Combobox
    * When `true`, the value-clearing will be disabled.
    */
   @Prop({ reflect: true }) clearDisabled = false;
+
+  /**
+   * Text for the component's filter input field.
+   */
+  @Prop({ reflect: true, mutable: true }) filterText = "";
+
+  @Watch("filterText")
+  filterTextChange(value: string): void {
+    this.updateActiveItemIndex(-1);
+    this.filterItems(value, true);
+  }
 
   /**
    * When `selectionMode` is `"ancestors"` or `"multiple"`, specifies the display of multiple `calcite-combobox-item` selections, where:
@@ -350,14 +361,14 @@ export class Combobox
 
     await componentOnReady(this.el);
 
-    if (!this.allowCustomValues && this.text) {
+    if (!this.allowCustomValues && this.filterText) {
       this.clearInputValue();
       this.filterItems("");
       this.updateActiveItemIndex(-1);
     }
 
-    if (this.allowCustomValues && this.text.trim().length) {
-      this.addCustomChip(this.text);
+    if (this.allowCustomValues && this.filterText.trim().length) {
+      this.addCustomChip(this.filterText);
     }
 
     this.open = false;
@@ -481,6 +492,7 @@ export class Combobox
     setUpLoadableComponent(this);
     this.updateItems();
     await setUpMessages(this);
+    this.filterItems(this.filterText, false, false);
   }
 
   componentDidLoad(): void {
@@ -517,9 +529,11 @@ export class Combobox
   //
   //--------------------------------------------------------------------------
 
+  @Element() el: HTMLCalciteComboboxElement;
+
   private allSelectedIndicatorChipEl: HTMLCalciteChipElement;
 
-  @Element() el: HTMLCalciteComboboxElement;
+  private filterTextMatchPattern: RegExp;
 
   placement: LogicalPlacement = defaultMenuPlacement;
 
@@ -550,14 +564,6 @@ export class Combobox
   @State() selectedHiddenChipsCount = 0;
 
   @State() selectedVisibleChipsCount = 0;
-
-  @State() text = "";
-
-  /** when search text is cleared, reset active to  */
-  @Watch("text")
-  textHandler(): void {
-    this.updateActiveItemIndex(-1);
-  }
 
   @State() effectiveLocale: string;
 
@@ -622,7 +628,7 @@ export class Combobox
 
   private clearInputValue(): void {
     this.textInput.value = "";
-    this.text = "";
+    this.filterText = "";
   }
 
   setFilteredPlacements = (): void => {
@@ -663,13 +669,13 @@ export class Combobox
       case "Tab":
         this.activeChipIndex = -1;
         this.activeItemIndex = -1;
-        if (this.allowCustomValues && this.text) {
-          this.addCustomChip(this.text, true);
+        if (this.allowCustomValues && this.filterText) {
+          this.addCustomChip(this.filterText, true);
           event.preventDefault();
         } else if (this.open) {
           this.open = false;
           event.preventDefault();
-        } else if (!this.allowCustomValues && this.text) {
+        } else if (!this.allowCustomValues && this.filterText) {
           this.clearInputValue();
           this.filterItems("");
           this.updateActiveItemIndex(-1);
@@ -760,8 +766,8 @@ export class Combobox
         } else if (this.activeChipIndex > -1) {
           this.removeActiveChip();
           event.preventDefault();
-        } else if (this.allowCustomValues && this.text) {
-          this.addCustomChip(this.text, true);
+        } else if (this.allowCustomValues && this.filterText) {
+          this.addCustomChip(this.filterText, true);
           event.preventDefault();
         } else if (!event.defaultPrevented) {
           if (submitForm(this)) {
@@ -780,7 +786,7 @@ export class Combobox
         if (this.activeChipIndex > -1) {
           event.preventDefault();
           this.removeActiveChip();
-        } else if (!this.text && this.isMulti()) {
+        } else if (!this.filterText && this.isMulti()) {
           event.preventDefault();
           this.removeLastChip();
         }
@@ -1060,11 +1066,7 @@ export class Combobox
 
   inputHandler = (event: Event): void => {
     const value = (event.target as HTMLInputElement).value;
-    this.text = value;
-    this.filterItems(value, true);
-    if (value) {
-      this.activeChipIndex = -1;
-    }
+    this.filterText = value;
   };
 
   getItemsAndGroups(): ComboboxChildElement[] {
@@ -1078,11 +1080,18 @@ export class Combobox
         isGroup(item) ? label === item.label : value === item.value && label === item.textLabel,
       );
 
-    return debounce((text: string, setOpenToEmptyState = false): void => {
+    return debounce((text: string, setOpenToEmptyState = false, emit = true): void => {
       const filteredData = filter(this.data, text);
       const itemsAndGroups = this.getItemsAndGroups();
 
+      const matchAll = text === "";
+
       itemsAndGroups.forEach((item) => {
+        if (matchAll) {
+          item.hidden = false;
+          return;
+        }
+
         const hidden = !find(item, filteredData);
         item.hidden = hidden;
         const [parent, grandparent] = item.ancestors;
@@ -1096,13 +1105,21 @@ export class Combobox
         }
       });
 
+      this.filterTextMatchPattern =
+        this.filterText && new RegExp(`(${escapeRegExp(this.filterText)})`, "i");
+
       this.filteredItems = this.getFilteredItems();
+      this.filteredItems.forEach((item) => {
+        item.filterTextMatchPattern = this.filterTextMatchPattern;
+      });
 
       if (setOpenToEmptyState) {
-        this.open = this.text.trim().length > 0 && this.filteredItems.length > 0;
+        this.open = this.filterText.trim().length > 0 && this.filteredItems.length > 0;
       }
 
-      this.calciteComboboxFilterChange.emit();
+      if (emit) {
+        this.calciteComboboxFilterChange.emit();
+      }
     }, 100);
   })();
 
@@ -1165,7 +1182,7 @@ export class Combobox
   }
 
   getFilteredItems(): HTMLCalciteComboboxItemElement[] {
-    return this.items.filter((item) => !item.hidden);
+    return this.filterText === "" ? this.items : this.items.filter((item) => !item.hidden);
   }
 
   private getSelectedItems = (): HTMLCalciteComboboxItemElement[] => {
@@ -1238,7 +1255,7 @@ export class Combobox
     if (this.textInput) {
       this.textInput.value = "";
     }
-    this.text = "";
+    this.filterText = "";
   }
 
   getItems(): HTMLCalciteComboboxItemElement[] {
@@ -1618,6 +1635,7 @@ export class Combobox
           role="combobox"
           tabindex={this.activeChipIndex === -1 ? 0 : -1}
           type="text"
+          value={this.filterText}
         />
       </span>
     );